+++ web/src/site/UpdateAvailableAlert.tsx
@@ -6,6 +6,9 @@ import { eventLogger } from '../tracking/eventLogger'
 const onClickCTA = () => {
     eventLogger.log('AlertUpdateAvailableCTAClicked')
 }
+const onClickChangelog = () => {
+    eventLogger.log('AlertUpdateAvailableChangelogClicked')
+}
 
 /**
  * A global alert telling the site admin that an updated version of the Sourcegraph
@@ -19,9 +22,17 @@ export const UpdateAvailableAlert: React.SFC<{
         partialStorageKey={`Update/${updateVersionAvailable}`}
         className={`alert-animated-bg alert-success update-available-alert ${className}`}
     >
-        <Icon className="icon-inline site-alert__link-icon" /> An update is available:&nbsp;
-        <a className="site-alert__link" href="https://about.sourcegraph.com" onClick={onClickCTA}>
+        <Icon className="icon-inline site-alert__link-icon mr-2" />An update is available:&nbsp;
+        <a className="site-alert__link" href="https://about.sourcegraph.com" target="_blank" onClick={onClickCTA}>
             <span className="underline">Sourcegraph Server {updateVersionAvailable}</span>
+        </a>&nbsp;-&nbsp;
+        <a
+            className="site-alert__link"
+            href="https://about.sourcegraph.com/changelog"
+            target="_blank"
+            onClick={onClickChangelog}
+        >
+            changelog
         </a>
     </DismissibleAlert>
 )